Restructure lib
