+++ premium_bond_checker/__init__.py
@@ -1,0 +1,1 @@
+from .client import Client
