+++ pyproject.toml
@@ -10,7 +10,7 @@ readme = "README.md"
 python = "^3.10"
 requests = "^2.31.0"
 pytz = ">=2024.2,<2026.0"
-holidays = ">=0.71,<=0.71"
+holidays = ">=0.71,<=1.0"
 
 [tool.poetry.group.dev.dependencies]
 responses = "^0.25.3"
