+++ tests/test_client.py
@@ -5,8 +5,9 @@ from datetime import date
 import responses
 from freezegun import freeze_time
 
-from premium_bond_checker.client import BondPeriod, CheckResult, Client, Result
+from premium_bond_checker import BondPeriod, Client
 from premium_bond_checker.exceptions import InvalidHolderNumberException
+from premium_bond_checker.models import CheckResult, Result
 
 
 def load_fixture(fixture_name: str) -> str:
